+++ das-jira-connector/src/main/java/com/rawlabs/das/jira/tables/DASJiraTable.java
@@ -125,42 +125,4 @@ public abstract class DASJiraTable implements DASTable {
     return OffsetDateTime.parse(dateString, formatter);
   }
 
-  // A helper to fallback to a default message if ever the body would be empty or null.
-  private static String mkMessage(String msg) {
-    if (msg == null || msg.trim().isEmpty()) {
-      return "Unknown JIRA API error";
-    }
-    return msg;
-  }
-
-  /**
-   * Helper method to create a DASSdk RuntimeException from an ApiException.
-   *
-   * <p>When the SDK exposes more exceptions (e.g. authentication errors), this method can
-   * investigate the ApiException and create the appropriate SDK exception.
-   *
-   * @param e the ApiException to convert
-   * @return the RuntimeException to throw
-   */
-  protected RuntimeException makeSdkException(ApiException e) {
-    return switch (e.getCode()) {
-      case 400 -> new DASSdkInvalidArgumentException(mkMessage(e.getResponseBody()), e);
-      case 401 -> new DASSdkUnauthenticatedException("unauthorized", e);
-      case 403 -> new DASSdkPermissionDeniedException("permission denied", e);
-      default ->
-          // For now, just create a generic SDK exception with the body of the ApiException
-          new DASSdkInvalidArgumentException(mkMessage(e.getResponseBody()), e);
-    };
-  }
-
-  protected RuntimeException makeSdkException(com.rawlabs.das.jira.rest.software.ApiException e) {
-    return switch (e.getCode()) {
-      case 400 -> new DASSdkInvalidArgumentException(mkMessage(e.getResponseBody()), e);
-      case 401 -> new DASSdkUnauthenticatedException("unauthorized", e);
-      case 403 -> new DASSdkPermissionDeniedException("permission denied", e);
-      default ->
-          // For now, just create a generic SDK exception with the body of the ApiException
-          new DASSdkInvalidArgumentException(mkMessage(e.getResponseBody()), e);
-    };
-  }
 }
